refactor: extraction d'ApiErrorPage (partition miroir-404 des layouts)
